+++ example/web_simple/blog.dart
@@ -20,7 +20,7 @@ main(){
     author.email = 'jorge@borges.com';
     author.age = 123;
     author.save();
-    return objectory[Author].find(where.sortBy('age'));
+    return objectory[Author].find(where.sortBy($Author.age));
   }).then((auths){
     print("===================================================================================");
     print(">> Authors ordered by age ascending");
@@ -40,7 +40,7 @@ main(){
     user.login = 'lsmith';
     user.email = 'lucy@smith.com';
     user.save();
-    return objectory[User].find(where.sortBy('login'));
+    return objectory[User].find(where.sortBy($User.login));
   }).then((usrs){
     print("===================================================================================");
     print(">> >> Users ordered by login ascending");